Move  aafi_enable_windows_VT100_output() to tools/common.c
